utils package refactoring
