+++ run_ibis_benchmark.py
@@ -8,10 +8,12 @@ import time
 import mysql.connector
 
 from report import DbReport
-from utils import (
+from utils_base_env import (
     find_free_port,
     KeyValueListParser,
     str_arg_to_bool,
+)
+from utils import (
     remove_fields_from_dict,
     convert_units,
 )
