+++ utils/__init__.py
@@ -2,18 +2,13 @@ from .utils import (
     check_fragments_size,
     check_support,
     cod,
-    combinate_requirements,
     compare_dataframes,
     convert_type_ibis2pandas,
-    execute_process,
     files_names_from_pattern,
-    find_free_port,
     import_pandas_into_module_namespace,
-    KeyValueListParser,
     load_data_pandas,
     mse,
     print_times,
-    str_arg_to_bool,
     split,
     timer_ms,
     print_results,
